+++ temba/middleware.py
@@ -1,4 +1,4 @@
-from __future__ import absolute_import, unicode_literals
+from __future__ import unicode_literals
 
 import cProfile
 import pstats
@@ -19,6 +19,19 @@ class ExceptionMiddleware(object):
             traceback.print_exc(exception)
 
         return None
+
+
+class OrgHeaderMiddleware(object):
+    """
+    Simple middleware to add a response header with the current org id, which can then be included in logs
+    """
+    def process_response(self, request, response):
+        # if we have a user, log our org id
+        if hasattr(request, 'user') and request.user.is_authenticated():
+            org = request.user.get_org()
+            if org:
+                response['X-Temba-Org'] = org.id
+        return response
 
 
 class BrandingMiddleware(object):
